fix: Improve check for installed clarinet
The mechanism used for checking if Clarinet is installed failed when a
path was returned from `getServer`, but the binary did not actually
exist. This happens on non-Windows machines, for which `getServer`
always just returns `clarinet`. When this happened, the result from
`spawnSync` contained an error which was not checked, later resulting in
the extension crashing when it tried to read the output from the
command, silently failing for the user.

Documentation is also added, clarifying the dependency on Clarinet and
linking to installation instructions.

Fixes: #57